@@ -6,7 +6,7 @@ namespace PTS\Hydrator;
 interface HydratorInterface
 {
 
-    public function hydrate(array $dto, string $class, array $rules);
+    public function hydrate(array $dto, string $class, array $rules): object;
 
     public function hydrateModel(array $dto, object $model, array $rules): void;
 }
